% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle3D_curv.R
\name{angle3D_curv}
\alias{angle3D_curv}
\title{angle3D function}
\usage{
angle3D_curv(A, B, C)
}
\arguments{
\item{A}{a \code{vector} corresponding to the x, y, and z coordinates of the first point forming the angle}

\item{B}{a \code{vector} corresponding to the x, y, and z coordinates of the middle point forming the angle}

\item{C}{a \code{vector} corresponding to the x, y, and z coordinates of the last point forming the angle}
}
\value{
a \code{vector} with the distance
}
\description{
compute the Euclidian distance between two points
}
