% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mean.r
\name{build_mean}
\alias{build_mean}
\title{The build_mean function build the mean variable, based on the computation of the mean}
\usage{
build_mean(Trials, data, item)
}
\arguments{
\item{Trials}{A vector of the length of the data list (number of analysed trials)}

\item{data}{A list with the trajectory data}

\item{item}{A character corresponding to the type of tracers we are refering to, either "F" for water or "A" for food}
}
\value{
A \code{list} with the column numbers with names containing the pattern
}
\description{
The build_mean function build the mean variable, based on the computation of the mean
}
\examples{
\dontrun{
build_mean(Trials, WF_tx, item="F")
}
}
\seealso{
\code{\link{compute_velocity}}
}
\author{
Pauline Provini \href{mailto:pauline.provini@gmail.com}{pauline.provini@gmail.com}
}
