% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_particle.R
\name{build_particle}
\alias{build_particle}
\title{build_particle function}
\usage{
build_particle(data, sequ)
}
\arguments{
\item{data}{a list corresponding to the Traj dataset for a given trial}

\item{sequ}{sequence corresponding to the column number corresponding to tx on the Traj dataset of a given trial}
}
\value{
A \code{list} of all the particles of a given trial, organizes in list of tx, ty, tz
}
\description{
Organize the data into Trials and Particle names with tx, ty, tz as columns
}
\author{
Pauline Provini \href{mailto:pauline.provini@gmail.com}{pauline.provini@gmail.com}
}
