% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_vel.r
\name{build_vel}
\alias{build_vel}
\title{The build_vel function build the velocity variable, based on the computation of the velocity}
\usage{
build_vel(Trials, data, duration)
}
\arguments{
\item{Trials}{A vector of the length of the data list (number of analysed trials)}

\item{data}{A list with the trajectory data}

\item{duration}{A number corresponding to the time lag between two data rows}
}
\value{
A \code{list} with the column numbers with names containing the pattern
}
\description{
The build_vel function build the velocity variable, based on the computation of the velocity
}
\seealso{
\code{\link{compute_velocity}}
}
\author{
Pauline Provini \href{mailto:pauline.provini@gmail.com}{pauline.provini@gmail.com}
}
