% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_curvature.R
\name{compute_curvature}
\alias{compute_curvature}
\title{compute_curvature function}
\usage{
compute_curvature(data, x, y, span)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{x}{a \code{string} of the variable name x}

\item{y}{a \code{string} of the variable name y}

\item{span}{a \code{numeric} corresponding to the span on which the curvature is computed}
}
\value{
A data.frame
}
\description{
Calculate the curvature (K : magnitude and orientation)
}
\author{
Pauline Provini \href{mailto:pauline.provini@gmail.com}{pauline.provini@gmail.com}
}
