% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_divergence.R
\name{compute_divergence}
\alias{compute_divergence}
\title{compute_divergence function}
\usage{
compute_divergence(A, B, C)
}
\arguments{
\item{A}{a \code{vector} corresponding to the x, y, and z coordinates of the first point}

\item{B}{a \code{vector} corresponding to the x, y, and z coordinates of the middle point}

\item{C}{a \code{vector} corresponding to the x, y, and z coordinates of the last point}
}
\value{
a \code{vector} with the angle in degrees and the curvature in m-1
}
\description{
compute the Angle and the Menger Curvature formed by three points
}
