% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_velocity.r
\name{compute_velocity}
\alias{compute_velocity}
\title{The compute_velocity function computes the velocity}
\usage{
compute_velocity(data, duration)
}
\arguments{
\item{data}{A list with the trajectory data}

\item{duration}{A number corresponding to the time lag between two data rows}
}
\value{
A \code{list} with the column numbers with names containing the pattern
}
\description{
The compute_velocity function computes the velocity
}
\author{
Pauline Provini \href{mailto:pauline.provini@gmail.com}{pauline.provini@gmail.com}
}
