% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_beginning.R
\name{delete_beginning}
\alias{delete_beginning}
\title{delete_beginning function}
\usage{
delete_beginning(data, trial)
}
\arguments{
\item{data}{a \code{data.frame} corresponding to FS_trial with at least
variables named "WF_tx", "WF_ty", "WF_tz","Grp"}

\item{trial}{an \code{integer} corresponding to the trial number to clean}
}
\value{
a \code{data.frame} with the cleaned data (same variables)
}
\description{
clean the particles (passed the entrance of the mouth)
}
