% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_curling.R
\name{delete_curling}
\alias{delete_curling}
\title{delete_curling function}
\usage{
delete_curling(data, Particles, trial)
}
\arguments{
\item{data}{a \code{data.frame} with variables named "WF_tx", "WF_ty", "WF_tz","Grp"}

\item{Particles}{a \code{vector} corresponding to the Particles names to clean}

\item{trial}{an \code{integer} corresponding to the trial number to clean}
}
\value{
a \code{data.frame} with the cleaned data (same variables)
}
\description{
clean the particles that circle back towards the entrance of
}
