% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{distance function}
\usage{
distance(df1, df2)
}
\arguments{
\item{df1}{a \code{data.frame} with x, y, and z components in this order}

\item{df2}{a \code{data.frame} with x, y, and z components in this order}
}
\value{
a \code{vector} with the distance
}
\description{
compute the Euclidian distance between two points
}
