% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{find_peaks}
\usage{
find_peaks(x, m = 3)
}
\arguments{
\item{x}{The vector where to find the peaks}

\item{m}{a number A (local) peak is defined as a point such that m points either side of it has a lower or equal value to it}
}
\value{
a vector
}
\description{
find peaks in a vector
}
\author{
Stasia Grinberg https://github.com/stas-g/findPeaks/blob/master/find_peaks.R
}
