% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sublist.R
\name{get_sublist}
\alias{get_sublist}
\title{get_sublist}
\usage{
get_sublist(lst, group_name)
}
\arguments{
\item{lst}{The \code{list} from which to extract the sublist}

\item{group_name}{\code{character} corresponding to the name of the variable we want to extract}
}
\value{
a \code{list}
}
\description{
get a list in a nested list, based on the name of the nested variable
}
\author{
Pauline Provini
}
