% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_df.R
\name{list_to_df}
\alias{list_to_df}
\title{list_to_df function}
\usage{
list_to_df(
  list,
  id.vars,
  replace_by = c("value.", ""),
  replace_L1 = "L1",
  replace_L2 = "L2"
)
}
\arguments{
\item{list}{a \code{list} to transform into a dataframe}

\item{id.vars}{a \code{vector} of id variables. Can be integer (variable position) or string (variable name). If blank, will use all non-measured variables.}

\item{replace_by}{a \code{vector} of 2 \code{string} the first item corresponding to the value to replace, the second item corresponding to the replacement string}

\item{replace_L1}{a \code{string} replacing the "L1" column name of the new dataframe}

\item{replace_L2}{a \code{string} replacing the "L2" column name of the new dataframe}
}
\value{
A \code{data.frame}
}
\description{
Transform a nested list to a dataframe
}
\seealso{
\code{\link[=melt]{melt()}} for argument details (L1, L2)
}
\author{
Pauline Provini \href{mailto:pauline.provini@gmail.com}{pauline.provini@gmail.com}
}
