% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Maya_Loc}
\alias{Maya_Loc}
\title{Maya_Loc}
\format{
A \code{list} of 10 lists corresponding to each analysed trial of carp and tilapia, containing
the locator data:
\describe{
\item{BonesLocators_relative}{"Locators" data}
\item{ope_tip_data.tx}{the x trajectory of the locator attached to the tip of the "opercula"}
\item{ope_tip_data.ty}{the y trajectory of the locator attached to the tip of the "opercula"}
\item{ope_tip_data.tz}{the z trajectory of the locator attached to the tip of the "opercula"}
\item{hyoid_tip_data.tx}{the x trajectory of the locator attached to the tip of the "hyoid"}
\item{hyoid_tip_data.ty}{the y trajectory of the locator attached to the tip of the "hyoid"}
\item{hyoid_tip_data.tz}{the z trajectory of the locator attached to the tip of the "hyoid"}.
\item{lower_jaw__data.tx}{the x trajectory of the locator attached to the tip of the "lower_jaw"}
\item{lower_jaw__data.ty}{the y trajectory of the locator attached to the tip of the "lower_jaw"}
\item{lower_jaw_data.tz}{the z trajectory of the locator attached to the tip of the "lower_jaw"}
\item{upper_jaw_tip_data.tx}{the x trajectory of the locator attached to the tip of the "upper_jaw"}
\item{upper_jaw_tip_data.ty}{the y trajectory of the locator attached to the tip of the "upper_jaw"}
\item{upper_jaw_tip_data.tz}{the z trajectory of the locator attached to the tip of the "upper_jaw"}
...
}
}
\usage{
Maya_Loc
}
\description{
A large nested list with the Locator (Loc) data exported from Maya
}
\keyword{datasets}
