% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Maya_RB}
\alias{Maya_RB}
\title{Maya_RB}
\format{
A \code{list} of 10 lists corresponding to each analysed trial of carp and tilapia, containing
the rigid body data:
\describe{
\item{hyoid}{"hyoid" rigid body data}
\item{hyoid.data.tx}{the x trajectory of the rigid body "hyoid"}
\item{hyoid.data.ty}{the y trajectory of the rigid body "hyoid"}
\item{hyoid.data.tz}{the z trajectory of the rigid body "hyoid"}
\item{hyoid.data.rx}{the x rotation of the rigid body "hyoid"}
\item{hyoid.data.ry}{the y rotation of the rigid body "hyoid"}
\item{hyoid.data.rz}{the z rotation of the rigid body "hyoid"}
\item{inf}{"lower jaw" rigid body data}
\item{sup}{"upper jaw" rigid body data}
\item{ope}{"operculum" rigid body data}
...
}
}
\usage{
Maya_RB
}
\description{
A large nested list with the rigid body (RB) data exported from Maya
}
\keyword{datasets}
