% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Maya_WF}
\alias{Maya_WF}
\title{Maya_WF}
\format{
A \code{list} of 13 lists corresponding to each analysed trial of carp and tilapia, containing
the water and food tracers data (F for fluid, A for food...)
\describe{
\item{frame}{frames of the trial}
\item{A01.data.tx}{the x trajectory of the food tracer "A01"}
\item{A01.data.ty}{the y trajectory of the food tracer "A01"}
\item{A01.data.tz}{the z trajectory of the food tracer "A01"}
\item{F01.data.tx}{the x trajectory of the water tracer "F01"}
\item{F01.data.ty}{the y trajectory of the water tracer "F01"}
\item{F01.data.tz}{the z trajectory of the water tracer "F01"}
}
}
\usage{
Maya_WF
}
\description{
A large nested list with the water and food (WF) data exported from Maya
}
\keyword{datasets}
