% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_particles.r
\name{mean_particles}
\alias{mean_particles}
\title{mean_particles function}
\usage{
mean_particles(trial, variable, item)
}
\arguments{
\item{trial}{string corresponding to the trial name}

\item{variable}{list containing the data for the variable of all the particles of a given variable (tx, ty, etc.)}

\item{item}{either F or A to choose what particle to average (F for fluid, A for food...)}
}
\value{
A \code{dataframe} with \code{mean}, \code{sd}, \code{min} and \code{max} of all the particles of a given variable (tx, ty, etc.)
}
\description{
Compute the mean, minimum, maximum and standard deviation of all the tracers
}
\seealso{
\code{\link{build_mean}}
}
\author{
Pauline Provini \href{mailto:pauline.provini@gmail.com}{pauline.provini@gmail.com}
}
