% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mean_ty_F}
\alias{Mean_ty_F}
\title{Mean_ty_F}
\format{
A \code{list} of 13 lists corresponding to each analysed trial of carp and tilapia, containing:
\describe{
\item{item}{character "F" for Food}
\item{average}{the mean of y trajectory over all the food tracers on each frame of the trial}
\item{std}{the standard deviation of y trajectory over all the food tracers on each frame of the trial}
\item{min}{the minimum of y trajectory over all the food tracers on each frame of the trial}
\item{max}{the maximum of y trajectory over all the food tracers on each frame of the trial}
}
}
\usage{
Mean_ty_F
}
\description{
A  nested list with the mean, minimum, maximum and standard deviation of ty data over the food tracers
}
\keyword{datasets}
