% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mean_tz_W}
\alias{Mean_tz_W}
\title{Mean_tz_W}
\format{
A \code{list} of 13 lists corresponding to each analysed trial of carp and tilapia, containing:
\describe{
\item{item}{character "F" for Water}
\item{average}{the mean of z trajectory over all the water tracers on each frame of the trial}
\item{std}{the standard deviation of z trajectory over all the water tracers on each frame of the trial}
\item{min}{the minimum of z trajectory over all the water tracers on each frame of the trial}
\item{max}{the maximum of z trajectory over all the water tracers on each frame of the trial}
}
}
\usage{
Mean_tz_W
}
\description{
A  nested list with the mean, minimum, maximum and standard deviation of tz data over the water tracers
}
\keyword{datasets}
