% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_nested.r
\name{name_nested}
\alias{name_nested}
\title{The name_nested function gives the same name to all the x_th columns of a nested list}
\usage{
name_nested(data, x, name)
}
\arguments{
\item{data}{A nested list}

\item{x}{An integer giving the position of the column to rename in the nested list}

\item{name}{A character}
}
\value{
A \code{list} with modified column names
}
\description{
The name_nested function gives the same name to all the x_th columns of a nested list
}
\examples{
\dontrun{
map(1:length(Time), ~names(x =Time[[.]][1]))
data <- name_nested(Time, 1 , "time")
map(1:length(data), ~names(x =data[[.]][1]))
}
}
\author{
Pauline Provini \href{mailto:pauline.provini@gmail.com}{pauline.provini@gmail.com}
}
