% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palette_IOFLOW}
\alias{palette_IOFLOW}
\title{palette_IOFLOW}
\format{
An object of class \code{character} of length 13.
}
\usage{
palette_IOFLOW
}
\description{
The color palette used to plot the figure of the IOFLOW analysis
#' @format A \code{vector} of length 13 with colors hex code
}
\keyword{datasets}
