% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Particle}
\alias{Particle}
\title{Particle}
\format{
A \code{list} of 13 lists corresponding to each analysed trial of carp and tilapia, containing
the trajectory of each particle analyzed in this trial :
\describe{
\item{A01}{Particle "A01"}
\item{A01.data.tx}{the x trajectory of the food tracer "A01"}
\item{A01.data.ty}{the y trajectory of the food tracer "A01"}
\item{A01.data.tz}{the z trajectory of the food tracer "A01"}
...
}
}
\usage{
Particle
}
\description{
A  nested list with all the Particle trajectory (tx, ty, and tz) for each Trial
}
\keyword{datasets}
