% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean.R
\name{plot_mean}
\alias{plot_mean}
\title{plot_mean}
\usage{
plot_mean(d, ylab = "", xlab = "", subtitle = "", color = c("black", p1[5]))
}
\arguments{
\item{d}{The dataframe returned by the mean_particles function (2 columns: mean and sd)}

\item{ylab}{a \code{string} with the name of the y axis}

\item{xlab}{a \code{string} with the name of the x axis}

\item{subtitle}{a \code{string} with the name of the subtitle}

\item{color}{a \code{vector} with the colors used in the plot}
}
\value{
A plot
}
\description{
plot_Mean function plots the mean and standard deviation of a given variable
}
