% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RB.R
\name{plot_RB}
\alias{plot_RB}
\title{plot_RB}
\usage{
plot_RB(trial, RB_name, component = c("tx", "ty", "tz", "rx", "ry", "rz"))
}
\arguments{
\item{trial}{The trial to plot}

\item{RB_name}{The name of the rigid body to plot}

\item{component}{The component to plot}
}
\value{
A plot
}
\description{
plot_RB function plots the traces of the Rigid Body data
}
