% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sequence}
\alias{Sequence}
\title{Sequence}
\format{
A \code{dataframe} of 15 rows  corresponding to each analysed trial of carp and tilapia and 5 columns:
\describe{
\item{name}{name of the trial}
\item{beg_FS}{beginning of the first strike}
\item{end_FS}{end of the first strike}
\item{beg_RF}{beginning of the first reverse flow} # tag:useless?
\item{end_EF}{end of the first reverse flow}
}
}
\usage{
Sequence
}
\description{
A dataframe with timing information gathered from the trial observation
}
\keyword{datasets}
