% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublist.r
\name{sublist}
\alias{sublist}
\title{sublist
Extract a list from a nested list}
\usage{
sublist(data, pattern)
}
\arguments{
\item{data}{A nested list}

\item{pattern}{A character}
}
\value{
A \code{list}
}
\description{
sublist
Extract a list from a nested list
}
\examples{
\dontrun{
Traj <- lapply(MayaWaterFood, subset, select=-1)
tx <- sublist(Traj, "tx")
}
}
\seealso{
\code{\link{to_nested_colnames}}
}
\author{
Pauline Provini \href{mailto:pauline.provini@gmail.com}{pauline.provini@gmail.com}
}
