% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data2.R
\name{tidy_data2}
\alias{tidy_data2}
\title{tidy_data2}
\usage{
tidy_data2(Trials, data1, data2)
}
\arguments{
\item{Trials}{A \code{vector} corresponding to a sequence from 1 to the number of trials}

\item{data1}{A \code{list} of the same length as Trials}

\item{data2}{A \code{list} of the same length as Trials}
}
\value{
A \code{list}
}
\description{
tidy_data2 combines two lists, keeping the names of nested lists
}
