% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Timing}
\alias{Timing}
\title{Timing}
\format{
A \code{list} of 13 lists corresponding to timing information:
\describe{
\item{frame}{frames of the trial}
\item{timing}{time of the trial (750 fps)}
\item{seq}{color code corresponding to phase (1st strike or reverse flow)}
}
}
\usage{
Timing
}
\description{
A  nested list with the timing data exported from Maya
}
\keyword{datasets}
