% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_nested_colnames.r
\name{to_nested_colnames}
\alias{to_nested_colnames}
\title{The to_nested_colnames function store the columns of a nested list, which names contain a given pattern}
\usage{
to_nested_colnames(data, pattern)
}
\arguments{
\item{data}{A nested list}

\item{pattern}{A character}
}
\value{
A \code{list} with the column numbers with names containing the pattern
}
\description{
The to_nested_colnames function store the columns of a nested list, which names contain a given pattern
}
\seealso{
\code{\link{sublist}}
}
\author{
Pauline Provini \href{mailto:pauline.provini@gmail.com}{pauline.provini@gmail.com}
}
