% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Trials}
\alias{Trials}
\title{Trials
A sequence number of the length of the analysed trials
#' @format A \code{vector} of length 13}
\format{
An object of class \code{integer} of length 13.
}
\usage{
Trials
}
\description{
Trials
A sequence number of the length of the analysed trials
#' @format A \code{vector} of length 13
}
\keyword{datasets}
