% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WF_tx}
\alias{WF_tx}
\title{WF_tx}
\format{
A \code{list} of 13 lists corresponding to each analysed trial of carp and tilapia, containing
the water and food tracers tx data (F for fluid, A for food...)
\describe{
\item{A01.data.tx}{the x trajectory of the food tracer "A01"}
}
}
\usage{
WF_tx
}
\description{
A  nested list with the water and food (WF) tx data exported from Maya
}
\keyword{datasets}
