% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WF_ty}
\alias{WF_ty}
\title{WF_ty}
\format{
A \code{list} of 13 lists corresponding to each analysed trial of carp and tilapia, containing
the water and food tracers ty data (F for fluid, A for food...)
\describe{
\item{A01.data.ty}{the y trajectory of the food tracer "A01"}
}
}
\usage{
WF_ty
}
\description{
A  nested list with the water and food (WF) ty data exported from Maya
}
\keyword{datasets}
