% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WF_tz}
\alias{WF_tz}
\title{WF_tz}
\format{
A \code{list} of 13 lists corresponding to each analysed trial of carp and tilapia, containing
the water and food tracers tz data (F for fluid, A for food...)
\describe{
\item{A01.data.tz}{the z trajectory of the food tracer "A01"}
}
}
\usage{
WF_tz
}
\description{
A  nested list with the water and food (WF) z data exported from Maya
}
\keyword{datasets}
