% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WF_vx}
\alias{WF_vx}
\title{WF_vx}
\format{
A \code{list} of 13 lists corresponding to each analysed trial of carp and tilapia, containing
the water and food tracers vx data (F for fluid, A for food...)
\describe{
\item{A01.data.vx}{the x velocity of the food tracer "A01"}
}
}
\usage{
WF_vx
}
\description{
A  nested list with the water and food (WF) vx data exported from Maya
}
\keyword{datasets}
