% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WF_vy}
\alias{WF_vy}
\title{WF_vy}
\format{
A \code{list} of 13 lists corresponding to each analysed trial of carp and tilapia, containing
the water and food tracers vy data (F for fluid, A for food...)
\describe{
\item{A01.data.vy}{the y velocity of the food tracer "A01"}
}
}
\usage{
WF_vy
}
\description{
A  nested list with the water and food (WF) vy data exported from Maya
}
\keyword{datasets}
