% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WF_vz}
\alias{WF_vz}
\title{WF_vz}
\format{
A \code{list} of 13 lists corresponding to each analysed trial of carp and tilapia, containing
the water and food tracers vz data (F for fluid, A for food...)
\describe{
\item{A01.data.vz}{the z velocity of the food tracer "A01"}
}
}
\usage{
WF_vz
}
\description{
A  nested list with the water and food (WF) vz data exported from Maya
}
\keyword{datasets}
